Support links even without contacts.
