+++ static/js/components/script_test.js
@@ -27,29 +27,4 @@ describe('script component', () => {
     let content = result.querySelector('div.call__script__body');
     expect(content.textContent).to.contain(issue.script);
   });
-
-  it('should NOT display script div if current contact does not exists', () => {
-    let cname = 'Senator Blowhart';
-    let id = 1;
-    let location = {params: {issueid: id}};
-    let issue = {
-      id: id, // does not match location issue id
-      name: 'Bozo the nominee',
-      reason: 'crazy',
-      script: 'Please vote against nominee Bozo'
-    };
-    let contact = {name: cname, party: 'Dem'};
-    issue.contacts = [contact];
-    let issues = [issue];
-    let contactIndices = {};
-    contactIndices[id] = 1; // contactIndex 1 does not exist
-    let state = {
-      issues,
-      location,
-      contactIndices
-    };
-    let result = script(state);
-    expect(result).to.be.undefined;
-  });
-
 });
